% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_methods.R
\docType{methods}
\name{append,color-method}
\alias{append,color-method}
\title{Appends two palettes to form a single palette.}
\usage{
\S4method{append}{color}(x, values, after = length(x))
}
\arguments{
\item{x}{the color palette to be modified.}

\item{values}{another color palette to be appended}

\item{after}{currently unimplemented.}
}
\description{
Applies to the \link[colorspace:color-class]{colorspace::color} class.
}
