% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_methods.R
\name{apply.color}
\alias{apply.color}
\title{Applies a function to each entry of a \link[colorspace:color-class]{colorspace::color} palette.}
\usage{
apply.color(X, FUN, ...)
}
\arguments{
\item{X}{the color palette}

\item{FUN}{the function to be applied}

\item{...}{extra arguments to FUN}
}
\value{
a list of each result of FUN applied to each entry in X
}
\description{
Applies a function to each entry of a \link[colorspace:color-class]{colorspace::color} palette.
}
