% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_methods.R
\docType{methods}
\name{as.list}
\alias{as.list}
\alias{as.list,color-method}
\title{Creates a list with single colors from a palette.}
\usage{
as.list(x, ...)

\S4method{as.list}{color}(x, ...)
}
\arguments{
\item{x}{color object to be coerced to a list}

\item{...}{ignored for this class}
}
\description{
Applies to the \link[colorspace:color-class]{colorspace::color} class.
}
\examples{
data("OptimGreenRedLAB")
as.list(OptimGreenRedLAB)
}
