% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_methods.R
\name{as}
\alias{as}
\title{Transforms palette to list of single colors.}
\description{
Applies to the \link[colorspace:color-class]{colorspace::color} class.
}
\examples{
data("OptimGreenRedLAB")
as(OptimGreenRedLAB, "list")
}
