% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{ColorPatchColorFun}
\alias{ColorPatchColorFun}
\title{Creates a color function mapping (ratio, conf) tuples to a single color}
\usage{
ColorPatchColorFun(palette = "OptimGreenRedLAB")
}
\arguments{
\item{palette}{name of the palette (see \code{\link[=data]{data()}}) - defaults to "OptimGreenRedLAB"}
}
\value{
A function mapping (ratio, conf) to a color.
}
\description{
Creates a color function mapping (ratio, conf) tuples to a single color
}
\examples{
fn <- ColorPatchColorFun("OptimBlueYelloLAB")
}
