% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{ColorPatchSizeFun}
\alias{ColorPatchSizeFun}
\title{Creates a size function mapping (ratio, conf) to a single color}
\usage{
ColorPatchSizeFun(type = "linear")
}
\arguments{
\item{type}{defaults to "linear"}
}
\value{
A function mapping (ratio, conf) to a size.
}
\description{
Creates a size function mapping (ratio, conf) to a single color
}
