% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{CreateClusteredData}
\alias{CreateClusteredData}
\title{Creates clustered random data}
\usage{
CreateClusteredData(nrow = 30, ncol = 12, nrow.clusters = 2,
  ncol.clusters = 2, alpha = 4)
}
\arguments{
\item{nrow}{Number of rows (default: 30)}

\item{ncol}{Number of columns (default: 12)}

\item{nrow.clusters}{Number of row cluster}

\item{ncol.clusters}{Number of column clusters (default: 2)}

\item{alpha}{Scaling facor (default: 4)}
}
\value{
A data set with $ratio and $conf values
}
\description{
Creates clustered random data
}
