% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{CreateExampleData}
\alias{CreateExampleData}
\title{Creates demonstration data of the colorpatch package}
\usage{
CreateExampleData(nrow = 30, ncol = 12)
}
\arguments{
\item{nrow}{number of rows (default 30)}

\item{ncol}{number of columns (default 12)}
}
\value{
the data set
}
\description{
Creates demonstration data of the colorpatch package
}
\examples{
library(ggplot2)
library(colorpatch)
dat <- CreateExampleData()
df <- ToDataFrame(dat)
p <- ggplot(df, aes(x = x, y = y, ratio = ratio, conf = conf)) 
p <- p + theme_colorpatch() + coord_fixed(ratio = 1)  + stat_colorpatch()
plot(p)
}
