% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{GeneratePalettes}
\alias{GeneratePalettes}
\title{Creates color palettes and saves them as files}
\usage{
GeneratePalettes(col.dist.fun = DistColorFun("LAB"), ...)
}
\arguments{
\item{col.dist.fun}{Color distance function.}

\item{...}{Additional arguments forwarded to \code{\link[colorpatch:OptimizeBiColor]{colorpatch::OptimizeBiColor()}}.}
}
\value{
Nothing - this function is used for its side effects (creating files in data).
}
\description{
Creates color palettes and saves them as files
}
