% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GreenRedRGB}
\alias{GreenRedRGB}
\title{Standard RGB Green/Red two-sided color scale.}
\format{An object of class \link[colorspace:color-class]{colorspace::color}.}
\usage{
GreenRedRGB
}
\description{
A two-sided color scale left side: green, center: black, right side: red.
}
\keyword{datasets}
