% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{HsvColorFun}
\alias{HsvColorFun}
\title{Creates a color function mapping ratio/conf values to a HSV colorspace}
\usage{
HsvColorFun(coerce.fn = colorspace::hex, hue.offset = 60, hue.scale = -60,
  saturation = 1)
}
\arguments{
\item{coerce.fn}{coerces each HSV color with this function (defaults \code{\link[colorspace:hex]{colorspace::hex()}})}

\item{hue.offset}{hue offset (defaults to 60)}

\item{hue.scale}{hue scale (defaults to 60)}

\item{saturation}{HSV saturation (defaults to 1)}
}
\value{
a color mapping function (ratio,conf) -> color
}
\description{
Creates a color function mapping ratio/conf values to a HSV colorspace
}
