% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{HsvSizeFun}
\alias{HsvSizeFun}
\title{Creates a size function mapping ratio/conf to a patch size for bicolorings}
\usage{
HsvSizeFun()
}
\value{
a size mapping function (ratio,conf) -> size
}
\description{
Creates a size function mapping ratio/conf to a patch size for bicolorings
}
