% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_methods.R
\docType{methods}
\name{length,color-method}
\alias{length,color-method}
\title{Returns the length of a palette (the number of entries).}
\usage{
\S4method{length}{color}(x)
}
\arguments{
\item{x}{an color object}
}
\description{
Applies to the \link[colorspace:color-class]{colorspace::color} class.
}
