% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OptimBlueYellowLAB}
\alias{OptimBlueYellowLAB}
\title{Optimum RGB Blue/Yellow two-sided color scale in LAB color space.}
\format{An object of class \link[colorspace:color-class]{colorspace::color}.}
\usage{
OptimBlueYellowLAB
}
\description{
A two-sided color scale left side: blue, center: black, right side: yellow.
}
\keyword{datasets}
