% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OptimGreenRedLAB}
\alias{OptimGreenRedLAB}
\title{Optimum RGB Green/Red two-sided color scale in LAB color space.}
\format{An object of class \link[colorspace:color-class]{colorspace::color}.}
\usage{
OptimGreenRedLAB
}
\description{
A two-sided color scale left side: green, center: black, right side: red.
}
\keyword{datasets}
