% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{OrderData}
\alias{OrderData}
\title{Orders rows and columns of data.}
\usage{
OrderData(dat, orderFn = OrderDataHclust, distFn = stats::dist)
}
\arguments{
\item{dat}{Ratio data}

\item{orderFn}{Ordering method (default: \link{OrderDataHclust})}

\item{distFn}{Distance function (Idefault \link[stats:dist]{stats::dist})}
}
\value{
ordered data
}
\description{
Orders rows and columns of data.
}
