% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{OrderDataHclust}
\alias{OrderDataHclust}
\title{Orders rows and column distances with \code{\link[stats:hclust]{stats::hclust()}}}
\usage{
OrderDataHclust(row.dist, col.dist, ...)
}
\arguments{
\item{row.dist}{row distances}

\item{col.dist}{column distances}

\item{...}{optional parameters forwarded to the \code{\link[stats:hclust]{stats::hclust()}} function}
}
\value{
a list with irow and icol containing the orders of rows and columns
}
\description{
Orders rows and column distances with \code{\link[stats:hclust]{stats::hclust()}}
}
