% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{OrderWithTSP}
\alias{OrderWithTSP}
\title{Orders a data set given a distance matrix with TSP}
\usage{
OrderWithTSP(dist, ...)
}
\arguments{
\item{dist}{distance object or distance matrix}

\item{...}{extra arguments fed to \code{\link[TSP:solve_TSP]{TSP::solve_TSP()}}}
}
\value{
a path (vector of integers)
}
\description{
Orders a data set given a distance matrix with TSP
}
