% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{PlotSymmetry}
\alias{PlotSymmetry}
\title{Plots the symmetry of a bivariate color scale}
\usage{
PlotSymmetry(pal, color.space = "LAB")
}
\arguments{
\item{pal}{A two-sided input palette \link[colorspace:color-class]{colorspace::color}}

\item{color.space}{Color space where the distances shall be computed (default "LAB")}
}
\value{
a ggplot object
}
\description{
Plots the symmetry of a bivariate color scale
}
\examples{
data("OptimGreenRedLAB")
PlotSymmetry(OptimGreenRedLAB)
}
