% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{PlotUniformity}
\alias{PlotUniformity}
\title{Plots the uniformity of a color palette}
\usage{
PlotUniformity(pal, color.space = "LAB")
}
\arguments{
\item{pal}{A colorspace palette}

\item{color.space}{the color space (see \link[colorspace:color-class]{colorspace::color})}
}
\value{
a ggplot instance
}
\description{
Plots the uniformity of a color palette
}
\examples{
data("OptimGreenRedLAB")
p <- PlotUniformity(OptimGreenRedLAB)
plot(p)
}
