% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{ReadArraySRGB}
\alias{ReadArraySRGB}
\title{Reads a sRGB color table as CSV file}
\usage{
ReadArraySRGB(file.name)
}
\arguments{
\item{file.name}{the color file}
}
\value{
a colorspace palette
}
\description{
Reads a sRGB color table as CSV file
}
