% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_ggplot.R
\docType{data}
\name{StatColorPatch}
\alias{StatColorPatch}
\title{A \link[ggplot2:ggproto]{ggplot2::ggproto} class for showing color patches.}
\format{An object of class \code{StatColorPatch} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\usage{
StatColorPatch
}
\description{
A \link[ggplot2:ggproto]{ggplot2::ggproto} class for showing color patches.
}
\keyword{datasets}
