% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_ggplot.R
\name{theme_colorpatch}
\alias{theme_colorpatch}
\title{A ggplot2 theme for rendering colorpatches (black background)}
\usage{
theme_colorpatch(fill = "black", plot.background = fill)
}
\arguments{
\item{fill}{background fill color (default: "black")}

\item{plot.background}{background fill color (default: "black")}
}
\value{
a theme function for showing color patches
}
\description{
A ggplot2 theme for rendering colorpatches (black background)
}
\examples{
library(ggplot2)
library(colorpatch)
dat <- CreateExampleData()
df <- ToDataFrame(dat)
p <- ggplot(df) + theme_colorpatch() + stat_colorpatch(aes(ratio=ratio,conf=conf,x=x,y=y))
}
