% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorpatch_impl.R
\name{ToDataFrame}
\alias{ToDataFrame}
\title{Transforms a ratio/conf data set to a ggplot dataframe}
\usage{
ToDataFrame(dat)
}
\arguments{
\item{dat}{must be a list with two matrices \code{ratio} and \code{conf}}
}
\value{
a data frame
}
\description{
Transforms a ratio/conf data set to a ggplot dataframe
}
